requirements.txt
antlr4-python3-runtime==4.13.0
antlr4-tools==0.2
pyinstaller==6.1.0
tk==0.1.0